// src/api/routes/department/types/department.types.ts
//: DepartmentDTO                                                          
/** 
 * @table `Department - Таблица с группами.`
 * @param id Id записи.
 * @param name Имя группы.
 * @param discription Описание группы.
 * @param icon Иконка группы, например: "16.png"
 */
export interface DepartmentDTO {
    id: number;
    name: string;
    discription: string;
    icon: string;
}

//: Types                                                                   